+++ cmd/frontend/internal/completions/BUILD.bazel
@@ -9,9 +9,9 @@ go_library(
         "//cmd/frontend/enterprise",
         "//cmd/frontend/internal/completions/resolvers",
         "//cmd/frontend/internal/guardrails",
+        "//cmd/frontend/internal/httpapi/completions",
         "//internal/codeintel",
         "//internal/cody",
-        "//internal/completions/httpapi",
         "//internal/conf/conftypes",
         "//internal/database",
         "//internal/observation",
