+++ cmd/frontend/internal/completions/resolvers/BUILD.bazel
@@ -7,9 +7,9 @@ go_library(
     visibility = ["//cmd/frontend:__subpackages__"],
     deps = [
         "//cmd/frontend/graphqlbackend",
+        "//cmd/frontend/internal/httpapi/completions",
         "//internal/cody",
         "//internal/completions/client",
-        "//internal/completions/httpapi",
         "//internal/completions/types",
         "//internal/conf",
         "//internal/database",
