+++ cmd/frontend/internal/httpapi/completions/codecompletion.go
@@ -1,4 +1,4 @@
-package httpapi
+package completions
 
 import (
 	"context"
