+++ cmd/frontend/internal/httpapi/completions/handler.go
@@ -1,4 +1,4 @@
-package httpapi
+package completions
 
 import (
 	"context"
