+++ cmd/frontend/internal/httpapi/completions/handler_test.go
@@ -1,4 +1,4 @@
-package httpapi
+package completions
 
 import (
 	"fmt"
